Cambia nombre de bundle.js a firebase.js
