+++ public/biblioteca.html
@@ -61,7 +61,7 @@
         crossorigin="anonymous"></script>
     <script src="js/biblioteca.js"></script>
     <!-- Firebase imports -->
-    <script src="js/bundle.js"></script>
+    <script src="js/firebase.js"></script>
 </body>
 
 </html>
