+++ public/calculadora.html
@@ -68,7 +68,7 @@
     <script src="js/calculadora.js"></script>
     <script src="js\loadNavFooter.js"></script>
     <!-- Firebase imports -->
-    <script src="js/bundle.js"></script>
+    <script src="js/firebase.js"></script>
 </body>
 
 </html>
