+++ public/contacto.html
@@ -65,7 +65,7 @@
         crossorigin="anonymous"></script>
     <script src="js\loadNavFooter.js"></script>
     <!-- Firebase imports -->
-    <script src="js/bundle.js"></script>
+    <script src="js/firebase.js"></script>
 </body>
 
 </html>
