+++ public/detallesProducto.html
@@ -189,7 +189,7 @@
     <script src="js/biblioteca.js"></script>
     <script src="js\loadNavFooter.js"></script>
     <!-- Firebase imports -->
-    <script src="js/bundle.js"></script>
+    <script src="js/firebase.js"></script>
 
 </body>
