+++ public/index.html
@@ -166,7 +166,7 @@
     <script src="js\loadNavFooter.js"></script>
     <script type="module" src="js/biblioteca.js"></script>
     <!-- Firebase imports -->
-    <script src="js/bundle.js"></script>
+    <script src="js/firebase.js"></script>
 </body>
 
 </html>
