--- src/index.js
@@ -1,17 +1,0 @@
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyDoDJzPNJG3AkGZkAvLBCaROpwDk2FYjVA",
-  authDomain: "pescar-3dseos.firebaseapp.com",
-  projectId: "pescar-3dseos",
-  storageBucket: "pescar-3dseos.appspot.com",
-  messagingSenderId: "185417562298",
-  appId: "1:185417562298:web:13cb3a79ac0af822ee6a83"
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
